Use shell grain as default in modules/cmdmod.py
This required moving the shell grain out of the core file so that
it could be imported into cmdmod. This was required because
__grains__ isn't part of the globals when cmdmod is first called
because grains/core imports it and for this same reason a function
in core couldn't be imported into cmdmod.
